# Reuse DCEL half edges (fix for geometry collection DCEL bugs)