--- a/geom/dcel_extract_intersection_matrix.go
+++ b/geom/dcel_extract_intersection_matrix.go
@@ -46,19 +46,19 @@ func (v *vertexRecord) location(operand operand) imLocation {
 	// NOTE: It's important that we check the Boundary flag before the Interior
 	// flag, since both might be set. In that case, we want to treat the
 	// location as a Boundary, since the boundary is a more specific case.
-	switch {
-	case v.locations[operand].boundary:
+	if v.locations[operand].boundary {
 		return imBoundary
-	case v.locations[operand].interior:
-		return imInterior
-	default:
-		// We don't know the location of the point. But it must be either
-		// Exterior or Interior because if it were Boundary, then we would know
-		// that. We can just use the location of one of the incident edges,
-		// since that would have the same location.
-		for _, e := range v.incidents {
-			return e.location(operand)
-		}
-		panic("point has no incidents") // Can't happen, due to ghost edges.
 	}
+	if v.locations[operand].interior {
+		return imInterior
+	}
+
+	// We don't know the location of the point. But it must be either Exterior
+	// or Interior because if it were Boundary, then we would know that due to
+	// an explicit flag. We can just use the location of one of the incident
+	// edges, since that would have the same location.
+	for e := range v.incidents {
+		return e.location(operand)
+	}
+	panic("point has no incidents") // Can't happen, due to ghost edges.
 }