… geom/alg_set_op_test.go
… geom/dcel_edge_set.go
… geom/dcel_test.go