## Description

commit c85bbacc35acfe974a783523e9c381cfab0a14a6
Author: Peter Stace <peterstace@gmail.com>
Date:   Fri Nov 11 11:21:29 2022 +1100

    Reuse DCEL half edges (fix for geometry collection DCEL bugs)
    
    The existing DCEL implementation has a bug where half-edge records are
    not reused if they appear in multiple polygons in the input geometry.
    Before the `simplefeatures` library supported GeometryCollection inputs,
    this bug didn't occur because the OGC Simple Feature Access spec
    prohibits MultiPolygons from self-touching along their boundaries.
    
    Similar bugs have existed in the past in scenarios where input
    LineStrings contain repeated or shared segments, or if the input
    geometries overlap with the "ghost" lines that join together separate
    parts of the input. Previously, the DCEL code handled cases via explicit
    checks and workarounds.
    
    This PR solves the same problem by using a more generic `addOrGetEdge`
    method that either gets an existing edge or creates a new one. The DCEL
    algorithm uses this new method to solve the existing repeated edge
    problem (previously solved via workaround) and the GeometryCollection
    bug.
    
    This approach results in considerably simpler code for loading
    geometries into a DCEL. It also means that the DCEL algorithm no longer
    needs an "overlay" step since it can load both geometries directly into
    the same DCEL structure. This is the opposite of the previous approach,
    where a DCEL for each operand was created separately and then merged
    together.


## Check List

Have you:

- Added unit tests? Yes.

- Add cmprefimpl tests? (if appropriate?) N/A, relies on existing.

- Updated release notes? (if appropriate?) N/A, relies on existing.

## Related Issue

- https://github.com/peterstace/simplefeatures/issues/465

- https://github.com/peterstace/simplefeatures/issues/300

## Benchmark Results

- There are some (relatively small) but unexpected (😄) performance improvements.

<details>

<summary>Click to expand</summary>

```
name                                                        old time/op    new time/op    delta
pkg:github.com/peterstace/simplefeatures/geom goos:linux goarch:arm64
LineEnvelope/0-4                                              2.07ns ± 1%    2.06ns ± 1%    -0.89%  (p=0.007 n=13+14)
LineEnvelope/1-4                                              2.10ns ± 2%    2.08ns ± 1%    -0.70%  (p=0.040 n=15+14)
LineEnvelope/2-4                                              2.08ns ± 2%    2.06ns ± 2%    -0.95%  (p=0.008 n=15+14)
LineEnvelope/3-4                                              2.09ns ± 2%    2.08ns ± 1%      ~     (p=0.359 n=15+10)
MarshalWKB/polygon/n=10-4                                     97.7ns ± 4%    97.1ns ± 2%      ~     (p=0.496 n=15+13)
MarshalWKB/polygon/n=100-4                                     282ns ± 5%     280ns ± 2%      ~     (p=0.397 n=13+13)
MarshalWKB/polygon/n=1000-4                                   1.92µs ± 6%    1.86µs ± 5%    -2.76%  (p=0.036 n=14+14)
MarshalWKB/polygon/n=10000-4                                  21.6µs ±34%    20.4µs ±22%      ~     (p=0.432 n=15+14)
UnmarshalWKB/polygon/n=10-4                                    159ns ± 3%     158ns ± 2%      ~     (p=0.511 n=13+14)
UnmarshalWKB/polygon/n=100-4                                   344ns ± 6%     344ns ± 2%      ~     (p=0.933 n=13+14)
UnmarshalWKB/polygon/n=1000-4                                 2.09µs ± 6%    2.11µs ± 9%      ~     (p=0.676 n=14+14)
UnmarshalWKB/polygon/n=10000-4                                21.0µs ±20%    21.2µs ±16%      ~     (p=0.769 n=14+14)
IntersectsLineStringWithLineString/n=10-4                      803ns ± 4%     809ns ± 3%      ~     (p=0.076 n=14+13)
IntersectsLineStringWithLineString/n=100-4                    12.1µs ± 2%    12.1µs ± 2%      ~     (p=0.940 n=14+12)
IntersectsLineStringWithLineString/n=1000-4                    133µs ± 2%     134µs ± 4%      ~     (p=0.252 n=14+15)
IntersectsLineStringWithLineString/n=10000-4                  2.20ms ±11%    2.30ms ±13%      ~     (p=0.098 n=15+15)
IntersectsMultiPointWithMultiPoint/n=20-4                      546ns ± 1%     547ns ± 2%      ~     (p=0.970 n=13+13)
IntersectsMultiPointWithMultiPoint/n=200-4                    7.38µs ± 2%    7.35µs ± 1%      ~     (p=0.181 n=14+13)
IntersectsMultiPointWithMultiPoint/n=2000-4                   70.0µs ± 1%    70.1µs ± 1%      ~     (p=0.480 n=14+13)
IntersectsMultiPointWithMultiPoint/n=20000-4                   765µs ± 2%     764µs ± 1%      ~     (p=0.880 n=13+13)
PolygonSingleRingValidation/n=10-4                            1.29µs ± 2%    1.29µs ± 4%      ~     (p=0.990 n=13+14)
PolygonSingleRingValidation/n=100-4                           19.1µs ± 2%    19.0µs ± 1%      ~     (p=0.641 n=14+13)
PolygonSingleRingValidation/n=1000-4                           276µs ± 1%     276µs ± 1%      ~     (p=0.494 n=14+12)
PolygonSingleRingValidation/n=10000-4                         3.75ms ± 8%    3.86ms ±10%      ~     (p=0.387 n=13+15)
PolygonMultipleRingsValidation/n=4-4                          3.67µs ± 1%    3.69µs ± 3%      ~     (p=0.519 n=13+13)
PolygonMultipleRingsValidation/n=36-4                         30.8µs ± 2%    30.9µs ± 2%      ~     (p=0.302 n=14+13)
PolygonMultipleRingsValidation/n=400-4                         405µs ± 2%     406µs ± 2%      ~     (p=0.982 n=14+14)
PolygonMultipleRingsValidation/n=4096-4                       5.06ms ± 2%    5.06ms ± 2%      ~     (p=0.946 n=14+14)
PolygonZigZagRingsValidation/n=10-4                           6.11µs ± 1%    6.17µs ± 2%    +0.96%  (p=0.015 n=14+14)
PolygonZigZagRingsValidation/n=100-4                          73.1µs ± 2%    72.9µs ± 1%      ~     (p=0.430 n=14+13)
PolygonZigZagRingsValidation/n=1000-4                          955µs ± 2%     952µs ± 1%      ~     (p=0.571 n=14+14)
PolygonZigZagRingsValidation/n=10000-4                        13.7ms ±15%    13.2ms ± 4%      ~     (p=0.354 n=15+14)
PolygonAnnulusValidation/n=10-4                               2.00µs ± 1%    2.02µs ± 2%    +0.81%  (p=0.015 n=14+14)
PolygonAnnulusValidation/n=100-4                              19.1µs ± 3%    19.1µs ± 3%      ~     (p=0.780 n=15+14)
PolygonAnnulusValidation/n=1000-4                              339µs ± 1%     342µs ± 1%    +0.93%  (p=0.005 n=14+13)
PolygonAnnulusValidation/n=10000-4                            4.50ms ± 3%    4.62ms ± 9%      ~     (p=0.156 n=13+15)
MultipolygonValidation/n=1-4                                   220ns ± 2%     221ns ± 2%      ~     (p=0.461 n=13+12)
MultipolygonValidation/n=4-4                                   493ns ± 1%     499ns ± 3%    +1.15%  (p=0.022 n=14+14)
MultipolygonValidation/n=16-4                                 2.17µs ± 2%    2.17µs ± 3%      ~     (p=0.821 n=15+14)
MultipolygonValidation/n=64-4                                 10.3µs ± 2%    10.3µs ± 2%      ~     (p=0.905 n=14+13)
MultipolygonValidation/n=256-4                                54.8µs ± 2%    54.8µs ± 2%      ~     (p=0.839 n=14+14)
MultipolygonValidation/n=1024-4                                336µs ± 3%     336µs ± 1%      ~     (p=0.762 n=13+13)
MultiPolygonTwoCircles/n=10-4                                 1.97µs ± 1%    1.99µs ± 3%      ~     (p=0.128 n=14+14)
MultiPolygonTwoCircles/n=100-4                                23.4µs ± 3%    23.4µs ± 2%      ~     (p=0.769 n=14+14)
MultiPolygonTwoCircles/n=1000-4                                269µs ± 1%     269µs ± 2%      ~     (p=0.689 n=13+12)
MultiPolygonTwoCircles/n=10000-4                              4.51ms ± 5%    4.56ms ± 9%      ~     (p=1.000 n=15+15)
MultiPolygonMultipleTouchingPoints/n=1-4                      2.59µs ± 2%    2.63µs ± 7%      ~     (p=0.626 n=14+14)
MultiPolygonMultipleTouchingPoints/n=10-4                     20.9µs ± 2%    20.8µs ± 2%      ~     (p=0.317 n=13+13)
MultiPolygonMultipleTouchingPoints/n=100-4                     260µs ± 3%     259µs ± 2%      ~     (p=0.172 n=15+14)
MultiPolygonMultipleTouchingPoints/n=1000-4                   3.26ms ± 1%    3.26ms ± 3%      ~     (p=0.259 n=14+13)
WKTParsing/point-4                                            1.12µs ± 2%    1.12µs ± 3%      ~     (p=0.245 n=14+12)
DistancePolygonToPolygonOrdering/n=100_swap=false-4           26.8µs ± 1%    26.8µs ± 2%      ~     (p=0.432 n=14+12)
DistancePolygonToPolygonOrdering/n=100_swap=true-4            26.8µs ± 1%    26.9µs ± 3%      ~     (p=0.756 n=14+13)
DistancePolygonToPolygonOrdering/n=1000_swap=false-4           468µs ± 5%     463µs ± 3%      ~     (p=0.525 n=15+13)
DistancePolygonToPolygonOrdering/n=1000_swap=true-4            458µs ± 1%     460µs ± 0%      ~     (p=0.183 n=14+11)
IntersectionPolygonWithPolygonOrdering/n=100_swap=false-4     3.23µs ± 3%    3.25µs ± 4%      ~     (p=0.510 n=13+14)
IntersectionPolygonWithPolygonOrdering/n=100_swap=true-4      3.25µs ± 3%    3.25µs ± 4%      ~     (p=0.830 n=14+13)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=false-4    34.9µs ± 3%    34.6µs ± 1%    -1.03%  (p=0.023 n=14+12)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=true-4     35.2µs ± 5%    34.7µs ± 1%      ~     (p=0.217 n=15+13)
MultiLineStringIsSimpleManyLineStrings/n=100-4                27.2µs ± 1%    27.1µs ± 2%      ~     (p=0.225 n=13+12)
MultiLineStringIsSimpleManyLineStrings/n=1000-4                327µs ± 2%     329µs ± 2%      ~     (p=0.150 n=14+14)
ForceCWandForceCCW/0-4                                        15.8ns ± 1%   38.7ns ±163%  +144.79%  (p=0.021 n=13+15)
ForceCWandForceCCW/0#01-4                                      100ns ± 3%      78ns ±80%      ~     (p=0.594 n=13+15)
ForceCWandForceCCW/1-4                                        16.0ns ± 2%   44.2ns ±128%  +176.22%  (p=0.011 n=12+15)
ForceCWandForceCCW/1#01-4                                      100ns ± 3%      73ns ±78%      ~     (p=0.080 n=11+15)
ForceCWandForceCCW/2-4                                        29.2ns ± 2%    29.2ns ± 1%      ~     (p=0.812 n=14+14)
ForceCWandForceCCW/2#01-4                                      154ns ± 2%     154ns ± 2%      ~     (p=0.729 n=13+14)
ForceCWandForceCCW/3-4                                        29.2ns ± 1%    29.4ns ± 2%      ~     (p=0.091 n=14+12)
ForceCWandForceCCW/3#01-4                                      155ns ± 2%     155ns ± 2%      ~     (p=0.285 n=14+11)
ForceCWandForceCCW/4-4                                        42.6ns ± 2%    42.6ns ± 1%      ~     (p=0.622 n=15+12)
ForceCWandForceCCW/4#01-4                                      239ns ± 2%     242ns ± 2%      ~     (p=0.090 n=13+12)
ForceCWandForceCCW/5-4                                        42.7ns ± 2%    42.8ns ± 1%      ~     (p=0.721 n=13+11)
ForceCWandForceCCW/5#01-4                                      241ns ± 5%     241ns ± 2%      ~     (p=0.340 n=12+13)
ForceCWandForceCCW/6-4                                        61.4ns ± 1%    61.6ns ± 1%      ~     (p=0.639 n=12+11)
ForceCWandForceCCW/6#01-4                                      386ns ± 3%     387ns ± 2%      ~     (p=0.577 n=12+11)
EnvelopeTransformXY-4                                          138ns ± 2%     137ns ± 1%    -0.87%  (p=0.007 n=15+12)
WKBParse/0-4                                                   412ns ± 1%     402ns ± 1%    -2.30%  (p=0.000 n=14+14)
WKBParse/1-4                                                   406ns ± 0%     407ns ± 1%      ~     (p=0.323 n=14+14)
WKBParse/2-4                                                   471ns ± 1%     471ns ± 1%      ~     (p=0.747 n=14+13)
WKBParse/3-4                                                   470ns ± 1%     471ns ± 1%      ~     (p=0.117 n=12+12)
WKBParse/4-4                                                   555ns ± 1%     555ns ± 1%      ~     (p=0.831 n=13+13)
WKBParse/5-4                                                   425ns ± 2%     424ns ± 1%      ~     (p=0.356 n=15+13)
WKBParse/6-4                                                   474ns ± 1%     473ns ± 1%      ~     (p=0.206 n=14+12)
WKBParse/7-4                                                   474ns ± 1%     476ns ± 1%    +0.46%  (p=0.014 n=14+13)
WKBParse/8-4                                                   555ns ± 1%     565ns ± 1%    +1.83%  (p=0.000 n=13+13)
WKBParse/9-4                                                   304ns ± 2%     304ns ± 1%      ~     (p=0.292 n=13+13)
WKBParse/10-4                                                  305ns ± 1%     306ns ± 1%      ~     (p=0.054 n=13+12)
WKBParse/11-4                                                  306ns ± 1%     307ns ± 3%      ~     (p=0.850 n=13+13)
WKBParse/12-4                                                  301ns ± 1%     302ns ± 2%      ~     (p=0.771 n=14+12)
WKBParse/13-4                                                  581ns ± 1%     593ns ± 2%    +1.98%  (p=0.000 n=13+13)
WKBParse/14-4                                                  686ns ± 0%     698ns ± 1%    +1.72%  (p=0.000 n=14+13)
WKBParse/15-4                                                  696ns ± 1%     687ns ± 1%    -1.28%  (p=0.000 n=14+12)
WKBParse/16-4                                                  820ns ± 1%     832ns ± 1%    +1.42%  (p=0.000 n=14+13)
WKBParse/17-4                                                  686ns ± 1%     695ns ± 1%    +1.33%  (p=0.000 n=14+13)
WKBParse/18-4                                                  888ns ± 1%     892ns ± 2%    +0.54%  (p=0.015 n=14+13)
WKBParse/19-4                                                  888ns ± 1%     879ns ± 1%    -0.99%  (p=0.000 n=14+12)
WKBParse/20-4                                                 1.08µs ± 1%    1.08µs ± 6%      ~     (p=0.173 n=14+14)
WKBParse/21-4                                                  294ns ± 1%     298ns ± 5%      ~     (p=0.083 n=15+14)
WKBParse/22-4                                                  296ns ± 1%     297ns ± 2%      ~     (p=0.272 n=15+12)
WKBParse/23-4                                                  296ns ± 1%     298ns ± 6%      ~     (p=0.901 n=15+13)
WKBParse/24-4                                                  292ns ± 1%     296ns ± 3%    +1.27%  (p=0.001 n=14+14)
WKBParse/25-4                                                 1.53µs ± 1%    1.56µs ± 6%      ~     (p=0.164 n=13+15)
WKBParse/26-4                                                 1.93µs ± 1%    1.94µs ± 1%      ~     (p=0.445 n=13+12)
WKBParse/27-4                                                 1.93µs ± 0%    1.94µs ± 1%    +0.28%  (p=0.003 n=9+12)
WKBParse/28-4                                                 2.43µs ± 1%    2.46µs ± 6%      ~     (p=0.742 n=13+15)
WKBParse/29-4                                                  294ns ± 1%     294ns ± 1%      ~     (p=1.000 n=13+12)
WKBParse/30-4                                                  295ns ± 1%     296ns ± 1%      ~     (p=0.155 n=13+12)
WKBParse/31-4                                                  296ns ± 1%     296ns ± 1%      ~     (p=0.810 n=14+12)
WKBParse/32-4                                                  291ns ± 1%     292ns ± 1%    +0.36%  (p=0.047 n=13+12)
WKBParse/33-4                                                  558ns ± 1%     562ns ± 4%      ~     (p=0.504 n=14+14)
WKBParse/34-4                                                  636ns ± 1%     639ns ± 2%      ~     (p=0.158 n=14+13)
WKBParse/35-4                                                  637ns ± 1%     648ns ± 1%    +1.69%  (p=0.000 n=15+13)
WKBParse/36-4                                                  698ns ± 1%     699ns ± 1%      ~     (p=0.291 n=14+12)
WKBParse/37-4                                                  799ns ± 1%     800ns ± 1%      ~     (p=0.640 n=14+12)
WKBParse/38-4                                                  938ns ± 1%     961ns ± 6%    +2.48%  (p=0.000 n=15+14)
WKBParse/39-4                                                  946ns ± 1%     953ns ± 3%      ~     (p=0.095 n=15+13)
WKBParse/40-4                                                 1.06µs ± 1%    1.06µs ± 3%      ~     (p=0.263 n=14+13)
WKBParse/41-4                                                  295ns ± 1%     296ns ± 1%    +0.49%  (p=0.032 n=14+13)
WKBParse/42-4                                                  296ns ± 0%     301ns ± 5%    +1.66%  (p=0.009 n=13+14)
WKBParse/43-4                                                  298ns ± 1%     298ns ± 1%      ~     (p=0.971 n=15+12)
WKBParse/44-4                                                  294ns ± 1%     294ns ± 1%      ~     (p=0.540 n=15+12)
WKBParse/45-4                                                  868ns ± 1%     872ns ± 1%    +0.45%  (p=0.024 n=15+12)
WKBParse/46-4                                                 1.06µs ± 1%    1.06µs ± 1%      ~     (p=1.000 n=14+12)
WKBParse/47-4                                                 1.05µs ± 1%    1.06µs ± 1%    +0.80%  (p=0.000 n=14+12)
WKBParse/48-4                                                 1.21µs ± 1%    1.21µs ± 1%      ~     (p=0.691 n=15+12)
WKBParse/49-4                                                 1.24µs ± 1%    1.24µs ± 1%      ~     (p=0.581 n=13+12)
WKBParse/50-4                                                 1.57µs ± 1%    1.57µs ± 1%      ~     (p=0.172 n=13+12)
WKBParse/51-4                                                 1.57µs ± 1%    1.57µs ± 1%    +0.36%  (p=0.037 n=15+12)
WKBParse/52-4                                                 1.84µs ± 1%    1.84µs ± 1%      ~     (p=0.271 n=15+12)
WKBParse/53-4                                                  295ns ± 1%     297ns ± 2%    +0.71%  (p=0.004 n=15+13)
WKBParse/54-4                                                  297ns ± 1%     300ns ± 3%    +0.89%  (p=0.007 n=14+13)
WKBParse/55-4                                                  298ns ± 1%     299ns ± 2%      ~     (p=0.289 n=15+13)
WKBParse/56-4                                                  294ns ± 1%     295ns ± 1%      ~     (p=0.105 n=15+12)
WKBParse/57-4                                                 1.88µs ± 0%    1.90µs ± 2%    +1.11%  (p=0.000 n=14+13)
WKBParse/58-4                                                 2.32µs ± 1%    2.34µs ± 3%      ~     (p=0.151 n=13+14)
WKBParse/59-4                                                 2.32µs ± 2%    2.33µs ± 2%    +0.60%  (p=0.012 n=14+14)
WKBParse/60-4                                                 2.81µs ± 1%    2.80µs ± 1%      ~     (p=0.268 n=14+12)
WKBParse/61-4                                                  294ns ± 1%     296ns ± 2%      ~     (p=0.215 n=14+13)
WKBParse/62-4                                                  296ns ± 1%     297ns ± 2%      ~     (p=0.929 n=14+12)
WKBParse/63-4                                                  295ns ± 1%     296ns ± 1%      ~     (p=0.291 n=14+12)
WKBParse/64-4                                                  292ns ± 1%     293ns ± 1%      ~     (p=0.051 n=15+12)
WKBParse/65-4                                                  840ns ± 1%     859ns ± 4%    +2.27%  (p=0.000 n=13+13)
WKBParse/66-4                                                 1.01µs ± 2%    1.00µs ± 2%    -0.66%  (p=0.031 n=15+13)
WKBParse/67-4                                                 1.00µs ± 1%    1.01µs ± 1%      ~     (p=0.282 n=15+12)
WKBParse/68-4                                                 1.11µs ± 1%    1.13µs ± 6%    +1.53%  (p=0.015 n=15+14)
pkg:github.com/peterstace/simplefeatures/geos goos:linux goarch:arm64
IntersectionWithoutValidation/n=10-4                          13.2µs ± 3%    13.3µs ± 2%      ~     (p=0.176 n=14+12)
IntersectionWithoutValidation/n=100-4                         28.7µs ± 3%    28.7µs ± 2%      ~     (p=0.910 n=14+14)
IntersectionWithoutValidation/n=1000-4                         146µs ± 4%     146µs ± 5%      ~     (p=0.838 n=15+15)
…